Support incremental levels.

+++ src/aim/common/util.h
@@ -201,7 +201,7 @@ static float GetStartSpeedForStopDistance(float stop_distance, float acceleratio
 }
 
 std::optional<std::string> StripLevelSuffix(const std::string& scenario_name,
-                                            int* level_out = nullptr);
+                                            float* level_out = nullptr);
 
 std::string AddLevelSuffix(const std::string& base_name, int level);
